requirements.txt
Django==2.2
django-storages==1.9.1
djangorestframework==3.11.0
